// src/types/RankedRoute.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RouteLeg } from "./RouteLeg";
import type { RouteMode } from "./RouteMode";
import type { Sustainability } from "./Sustainability";

export type RankedRoute = { mode: RouteMode, legs: Array<RouteLeg>, cr_per_hour: bigint, profit_per_cycle: bigint, cycle_seconds: number, total_jumps: number, sustainability: Sustainability, score: number, freshest_age_seconds: number, touches_fleet_carrier: boolean, route_hash: string, };