// src/types/RouteFilter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RouteMode } from "./RouteMode";

export type RouteFilter = { modes: Array<RouteMode> | null, max_jumps: number | null, min_cr_per_hour: bigint | null, max_profit_per_ton: number | null, pad_size_min: string | null, allow_anarchy: boolean, require_fleet_carrier: boolean, exclude_fleet_carrier: boolean, limit: number, };